Update tiramisu-savings-club2.js
So, I have made changes to the "should increase group balance by message value when the valid amount is provided" test case, and it is working and passing now.

Instead of trying to directly access the "groups" Struct variable, we need to use Smart Contract's getGroup(_groupId) method to get the "group" variable, and then use the member variables of the Struct.

There was a missing parenthesis "})" at the end of the file. Added that as well.